shadow-usdclp/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pydantic==2.10.3
numpy==2.1.3
pandas==2.2.3
scipy==1.14.1
statsmodels==0.14.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
